requirements.txt
# Core Libraries
numpy
pandas
scipy
scikit-learn

# Visualization
matplotlib
seaborn

# Deep Learning Frameworks
tensorflow
keras
torch
torchvision

# Natural Language Processing (NLP)
nltk
gensim
spaCy

# Computer Vision
opencv-python
Pillow

# Feature Engineering
featuretools

# Data Visualization
plotly

# Jupyter Notebook (if using Jupyter)
jupyter

# Miscellaneous
joblib
pyyaml
tqdm
fastapi
python-multipart